+++ changelog.md
@@ -6,8 +6,9 @@
 #### 10.0.6 [Dec 2022]
 - **[fix]** - extra dots in interwiki links #510
 - **[new]** - configure unsupported template behaviour - templateFallbackFn #509
-- **[fix]** - embedded infobox issue #506
 - **[update]** - dependencies
+<!-- embedded infobox change -->
+<!-- rollback interwiki change -->
 
 #### 10.0.5 [Dec 2022]
 - **[fix]** - broken cli script #504